feat: implement logo downloader handlers at specefic file to enhance code readability
